+++ lib/coal.dart
@@ -1,3 +1,4 @@
 export 'args.dart';
+export 'keypass.dart';
 export 'tab.dart';
 export 'utils.dart';