+++ lib/keypass.dart
@@ -1,0 +1,4 @@
+export 'src/keypass/decoder.dart';
+export 'src/keypass/dispatcher.dart';
+export 'src/keypass/event.dart';
+export 'src/keypass/keypass.dart';